+++ .gitignore
@@ -4,4 +4,3 @@ coverage/
 node_modules/
 .yarn.lock
 package-lock.json
-firebase-debug.log
