+++ src/index.html
@@ -35,9 +35,6 @@
     firebase.initializeApp(firebaseConfig);
   </script>
 
-
-
-
 <script type="module" src="main.js"></script>
 </body>
