+++ src/pages/feed/index.js
@@ -1,0 +1,14 @@
+import profileComponent from "../../components/profile/profile.js";
+
+const createHTML = () => {
+  return `
+    <h1>Feed</h1>
+    ${profileComponent.createHTML()}
+  `;
+};
+
+const registerListeners = () => {
+  profileComponent.registerListeners();
+};
+
+export default { createHTML, registerListeners };
